Trying to merge

+++ Build/clean.py
@@ -1,5 +1,6 @@
 import os
 import shutil
+import glob
 
 build_dir = "."
 
@@ -11,7 +12,12 @@ for file in filtered_files:
     os.remove(full_path)
 
 intermediate_dir = "./Intermediate"
-shutil.rmtree(intermediate_dir)
-    
+if os.path.exists(intermediate_dir):
+    shutil.rmtree(intermediate_dir)
+
+files = glob.glob("../Chapters/*/*/draft.pdf")
+for draft_file in files:
+    os.remove(draft_file)
+
 print('Cleaned.')
